VM: Allow integers of all sizes in `IntConst`.
- Large integers now don't have to take a spot in the constants table, as the instruction is wide enough to keep integers of all sizes.
- The backend instructions are split into an `IntConst` and `IntConst64` as `IntConst` for integers that fit into 16 bits has less memory to access.
- A corresponding `RealConst` instruction would not be advisable as long as Reals are allocated on the heap. A `RealConst` instruction would need to reallocate a new Real value each time it's called, while a `Const` instruction with a constant Real value does not need any allocations.

+++ compiler/src/lore/compiler/poem/PoemInstruction.scala
@@ -63,7 +63,7 @@ object PoemInstruction {
 
   case class Const(target: PReg, value: PVal) extends PoemInstruction(PoemOperation.Const)
 
-  case class IntConst(target: PReg, value: Int) extends PoemInstruction(PoemOperation.IntConst)
+  case class IntConst(target: PReg, value: Long) extends PoemInstruction(PoemOperation.IntConst)
   case class IntToReal(target: PReg, value: PReg) extends PoemInstruction(PoemOperation.IntToReal)
 
   case class BooleanConst(target: PReg, value: Boolean) extends PoemInstruction(PoemOperation.BooleanConst)
